refactor: scanner cooldown
