+++ server/src/services/functions/getServerSettings.js
@@ -12,6 +12,7 @@ function getServerSettings(req) {
   const user = {
     ...(req.user ? req.user : req.session),
     loggedIn: !!req.user,
+    cooldown: req.session?.cooldown || 0,
   }
 
   const { clientValues, clientMenus } = clientOptions(user.perms)
