+++ src/components/layout/dialogs/scanner/scanZone/PopupContent.jsx
@@ -5,10 +5,10 @@ import { useTranslation } from 'react-i18next'
 import debounce from 'lodash.debounce'
 
 import AdvancedAccordion from '@components/layout/custom/AdvancedAccordion'
-import { useScanStore } from '@hooks/useStore'
 
 import { StyledSubHeader } from '../Shared'
 import { ConfigContext } from '../ContextProvider'
+import { useScanStore } from '../store'
 
 const RADIUS_CHOICES = /** @type {const} */ (['pokemon', 'gym'])
 
@@ -54,7 +54,7 @@ export function ScanZonePopup() {
 /**
  *
  * @param {{
- *  name: keyof import("@rm/types").OnlyType<import('@hooks/useStore').UseScanStore, number>,
+ *  name: keyof import("@rm/types").OnlyType<import('../store').UseScanStore, number>,
  * } & import('@mui/material').SliderProps} props
  * @returns
  */
